refactor(lang_specs): remove unnecessary `styles` directory
