+++ lua/codedocs/lang_specs/bash/comment.lua
