+++ lua/codedocs/lang_specs/bash/func.lua
