+++ lua/codedocs/lang_specs/bash/init.lua
@@ -65,8 +65,8 @@ return {
 	styles = {
 		default = "Google",
 		definitions = {
-			func = require "codedocs.lang_specs.bash.styles.func",
-			comment = require "codedocs.lang_specs.bash.styles.comment",
+			func = require "codedocs.lang_specs.bash.func",
+			comment = require "codedocs.lang_specs.bash.comment",
 		},
 	},
 	extraction = {
