--- lua/codedocs/lang_specs/bash/styles/init.lua
@@ -1,1 +1,0 @@
-return
