+++ lua/codedocs/lang_specs/go/func.lua
