+++ lua/codedocs/lang_specs/javascript/class.lua
