+++ lua/codedocs/lang_specs/javascript/comment.lua
