+++ lua/codedocs/lang_specs/javascript/func.lua
