+++ lua/codedocs/lang_specs/javascript/init.lua
@@ -97,9 +97,9 @@ return {
 	styles = {
 		default = "JSDoc",
 		definitions = {
-			func = require "codedocs.lang_specs.javascript.styles.func",
-			class = require "codedocs.lang_specs.javascript.styles.class",
-			comment = require "codedocs.lang_specs.javascript.styles.comment",
+			func = require "codedocs.lang_specs.javascript.func",
+			class = require "codedocs.lang_specs.javascript.class",
+			comment = require "codedocs.lang_specs.javascript.comment",
 		},
 	},
 	extraction = {
