+++ lua/codedocs/lang_specs/kotlin/init.lua
@@ -82,9 +82,9 @@ return {
 	styles = {
 		default = "KDoc",
 		definitions = {
-			func = require "codedocs.lang_specs.kotlin.styles.func",
-			class = require "codedocs.lang_specs.kotlin.styles.class",
-			comment = require "codedocs.lang_specs.kotlin.styles.comment",
+			func = require "codedocs.lang_specs.kotlin.func",
+			class = require "codedocs.lang_specs.kotlin.class",
+			comment = require "codedocs.lang_specs.kotlin.comment",
 		},
 	},
 	extraction = {
