+++ lua/codedocs/lang_specs/php/func.lua
