+++ lua/codedocs/lang_specs/python/class.lua
