+++ lua/codedocs/lang_specs/typescript/class.lua
