+++ lua/codedocs/lang_specs/typescript/comment.lua
