+++ lua/codedocs/lang_specs/typescript/func.lua
